pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["tad hensley tad@novumhomeloans.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
flask = "^3.0.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
google-api-python-client = "^2.32.0"
google-auth = "^2.20.0"
google-auth-httplib2 = ">=0.1.0"
google-auth-oauthlib = "^1.0.0"
replit = "^4.1.0"
schedule = "^1.2.2"
xhtml2pdf = "^0.2.11"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[nix]
pkgs = [
  "pkgs.postgresql"
]
